+++ R/dummy_template.R
@@ -1,0 +1,10 @@
+#' Dummy template.
+#'
+#' This template creates a dummy output.
+#'
+#' @include chevron_tlg-S4class.R
+#' @export
+#'
+#' @examples
+#' run(dummy_template, syn_data)
+dummy_template <- chevron_simple()