+++ tests/testthat/test-chevron_tlg-class.R
@@ -160,3 +160,22 @@ test_that("chevron_g constructor returns an error when expected", {
     fixed = TRUE
   )
 })
+
+
+# chevron_simple ----
+
+test_that("chevron_simple works correctly", {
+  obj <- chevron_simple()
+  expect_silent(main(obj) <- \(adam_db, ...) abc)
+  expect_silent(main(obj) <- (\(adam_db, ...) {
+    abc
+  }))
+})
+
+test_that("chevron_simple errors if contains return", {
+  obj <- chevron_simple()
+  expect_error(
+    main(obj) <- \(adam_db, ...) return(abc),
+    "Must be a simple expression without `return`."
+  )
+})