+++ NEWS.md
@@ -1,6 +1,8 @@
 # chevron 0.2.4.9002
 
 * `MNG01` plot can now be displayed without error bars and can display a continuous temporal scale on the `x` axis.
+* Add a `chevron_simple` class only contain main function.
+* Remove `details` argument in `script_funs`, add `name` argument.
 
 # chevron 0.2.4